// server/src/handlebars-controllers/reviews/review.controller.interface.ts
import {IControllerRoute} from "../../common/route.interface";
import {NextFunction, Request, Response, Router} from "express";

export interface IReviewController {
    router: Router;
    bindRoutes: (routes: IControllerRoute[])=> void;
    addReview: (req: Request, res: Response, next: NextFunction)=> void;
    getAllReviews: (req: Request, res: Response, next: NextFunction)=> void;
}